refactor(visualization): extract dataMocks from util into its own mocks/ root folder
dataMocks is a ~2500-line cross-layer test-fixture file that was sitting in
util/ and obscuring real util problems. Move it to app/codeCharta/mocks/ (a
sibling of util/), repoint all 57 importers, and fix its one same-dir import
(./codeMapHelper -> ../util/codeMapHelper). Pure relocation: dependency-cruiser
0 errors, tests green across every import depth.
