+++ visualization/app/codeCharta/features/3dPrint/components/export3DMapDialog/export3DMapDialog.component.spec.ts
@@ -2,7 +2,7 @@ import { State } from "@ngrx/store"
 import { render, screen } from "@testing-library/angular"
 import userEvent from "@testing-library/user-event"
 import { AmbientLight, BufferGeometry, DirectionalLight, Group, Scene, Shape, Vector2, WebGLRenderer } from "three"
-import { DEFAULT_SETTINGS, DEFAULT_STATE, FILE_META, TEST_NODES } from "../../../../util/dataMocks"
+import { DEFAULT_SETTINGS, DEFAULT_STATE, FILE_META, TEST_NODES } from "../../../../mocks/dataMocks"
 import { CodeMapMesh } from "../../../../features/codeMap/facade"
 import { ThreeSceneService } from "../../../../features/codeMap/facade"
 import { Export3DMapDialogComponent } from "./export3DMapDialog.component"
