+++ visualization/app/codeCharta/features/codeMap/rendering/codeMapBuilding.spec.ts
@@ -1,5 +1,5 @@
 import { CodeMapBuilding } from "./codeMapBuilding"
-import { TEST_NODE_ROOT } from "../../../util/dataMocks"
+import { TEST_NODE_ROOT } from "../../../mocks/dataMocks"
 import { Vector3 } from "three"
 import { ColorConverter } from "../../../util/color/colorConverter"
