+++ visualization/app/codeCharta/features/codeMap/threeViewer/threeSceneService.spec.ts
@@ -6,7 +6,7 @@ import {
     TEST_NODES,
     VALID_FILE_NODE_WITH_ID,
     VALID_NODES_WITH_ID
-} from "../../../util/dataMocks"
+} from "../../../mocks/dataMocks"
 import { CODE_MAP_BUILDING, CODE_MAP_BUILDING_TS_NODE, CONSTANT_HIGHLIGHT } from "../rendering/codeMapBuilding.mocks"
 import { CodeMapBuilding } from "../rendering/codeMapBuilding"
 import { ThreeSceneService } from "./threeSceneService"
