+++ visualization/app/codeCharta/features/fileExtensionBar/selectors/hoveredNodeMetricDistribution.selector.spec.ts
@@ -1,6 +1,6 @@
 import { hoveredNodeMetricDistributionSelector } from "./hoveredNodeMetricDistribution.selector"
 import { FileExtensionCalculator } from "../../../util/fileExtension/fileExtensionCalculator"
-import { VALID_NODE_WITH_MULTIPLE_FOLDERS, VALID_NODE_WITH_MULTIPLE_FOLDERS_REVERSED } from "../../../util/dataMocks"
+import { VALID_NODE_WITH_MULTIPLE_FOLDERS, VALID_NODE_WITH_MULTIPLE_FOLDERS_REVERSED } from "../../../mocks/dataMocks"
 import { NodeType } from "../../../codeCharta.model"
 
 describe("hoveredNodeMetricDistributionSelector", () => {
