+++ visualization/app/codeCharta/features/labelSettings/components/labelSettingsPanel/labelSettingsPanel.component.spec.ts
@@ -6,7 +6,7 @@ import { setLabelMode } from "../../../../state/store/appSettings/labelMode/labe
 import { setLabelSize } from "../../../../state/store/appSettings/labelSize/labelSize.actions"
 import { setLabelsPerMap } from "../../../../state/store/appSettings/labelsPerMap/labelsPerMap.actions"
 import { setDelta, setFiles } from "../../../../state/store/files/files.actions"
-import { FILE_STATES_TWO_FILES, TEST_FILE_DATA, TEST_FILE_DATA_JAVA } from "../../../../util/dataMocks"
+import { FILE_STATES_TWO_FILES, TEST_FILE_DATA, TEST_FILE_DATA_JAVA } from "../../../../mocks/dataMocks"
 import { fireEvent } from "@testing-library/angular"
 import { LabelSettingsPanelComponent } from "./labelSettingsPanel.component"
 import { Store, StoreModule } from "@ngrx/store"
