+++ visualization/app/codeCharta/features/labelSettings/services/labelCollision.service.spec.ts
@@ -14,7 +14,7 @@ import { StateAccessStore } from "../stores/stateAccess.store"
 import { setHeightMetric } from "../../../state/store/dynamicSettings/heightMetric/heightMetric.actions"
 import { setLabelsPerMap } from "../../../state/store/appSettings/labelsPerMap/labelsPerMap.actions"
 import { setFiles } from "../../../state/store/files/files.actions"
-import { FILE_STATES_TWO_FILES } from "../../../util/dataMocks"
+import { FILE_STATES_TWO_FILES } from "../../../mocks/dataMocks"
 
 describe("LabelCollisionService", () => {
     let store: Store<CcState>
