+++ visualization/app/codeCharta/features/navBar/services/fileSelectionMode.service.spec.ts
@@ -2,7 +2,7 @@ import { TestBed } from "@angular/core/testing"
 import { FileSelectionState } from "../../../model/files/files"
 import { referenceFileSelector } from "../../../state/selectors/referenceFile/referenceFile.selector"
 import { addFile, removeFiles, setDelta, setStandard } from "../../../state/store/files/files.actions"
-import { TEST_FILE_DATA, TEST_FILE_DATA_JAVA } from "../../../util/dataMocks"
+import { TEST_FILE_DATA, TEST_FILE_DATA_JAVA } from "../../../mocks/dataMocks"
 import { FileSelectionModeService } from "./fileSelectionMode.service"
 import { State, Store, StoreModule } from "@ngrx/store"
 import { appReducers, setStateMiddleware } from "../../../state/store/state.manager"
