+++ visualization/app/codeCharta/features/navBar/services/uploadFiles.service.spec.ts
@@ -4,7 +4,7 @@ import { LoadFileService } from "../../../features/loadFile/facade"
 import { setIsLoadingFile } from "../../../state/store/appSettings/isLoadingFile/isLoadingFile.actions"
 import { setIsLoadingMap } from "../../../state/store/appSettings/isLoadingMap/isLoadingMap.actions"
 import { createCCFileInput } from "./createCCFileInput"
-import { TEST_FILE_CONTENT } from "../../../util/dataMocks"
+import { TEST_FILE_CONTENT } from "../../../mocks/dataMocks"
 import stringify from "safe-stable-stringify"
 import { EffectsModule } from "@ngrx/effects"
 import { Store, StoreModule } from "@ngrx/store"
