+++ visualization/app/codeCharta/features/sidebarExplorer/selectors/sortNode.spec.ts
@@ -6,7 +6,7 @@ import {
     VALID_NODE_WITH_MULTIPLE_FOLDERS_REVERSED,
     VALID_NODE_WITH_MULTIPLE_FOLDERS_SORTED_BY_NAME,
     VALID_NODE_WITH_MULTIPLE_FOLDERS_SORTED_BY_UNARY
-} from "../../../util/dataMocks"
+} from "../../../mocks/dataMocks"
 import { CodeMapNode, NodeType, SortingOption } from "../../../codeCharta.model"
 import { sortNodesInPlace } from "./sortNodesInPlace"
