+++ visualization/app/codeCharta/features/viewCubeToolbox/components/viewCubeToolbox/viewCubeToolbox.component.spec.ts
@@ -6,7 +6,7 @@ import { of } from "rxjs"
 import { metricDataSelector } from "../../../../state/selectors/accumulatedData/metricData/metricData.selector"
 import { defaultState } from "../../../../state/store/state.manager"
 import { ThreeMapControlsService } from "../../../../features/codeMap/facade"
-import { METRIC_DATA } from "../../../../util/dataMocks"
+import { METRIC_DATA } from "../../../../mocks/dataMocks"
 import { GlobalSettingsFacade } from "../../../globalSettings/facade"
 import { ScreenshotService } from "../../services/screenshot.service"
 import { ViewCubeToolboxComponent } from "./viewCubeToolbox.component"
