+++ visualization/app/codeCharta/state/effects/updateFileSettings/utils/blacklist.merger.spec.ts
@@ -1,5 +1,5 @@
 import { getMergedBlacklist } from "./blacklist.merger"
-import { TEST_FILE_DATA } from "../../../../util/dataMocks"
+import { TEST_FILE_DATA } from "../../../../mocks/dataMocks"
 import { BlacklistItem, CCFile } from "../../../../codeCharta.model"
 import { clone } from "../../../../util/clone"
