+++ visualization/app/codeCharta/state/effects/updateFileSettings/utils/edges.merger.spec.ts
@@ -1,5 +1,5 @@
 import { getMergedEdges } from "./edges.merger"
-import { TEST_FILE_DATA } from "../../../../util/dataMocks"
+import { TEST_FILE_DATA } from "../../../../mocks/dataMocks"
 import { CCFile, Edge } from "../../../../codeCharta.model"
 import { clone } from "../../../../util/clone"
