+++ visualization/app/codeCharta/state/effects/updateQueryParameters/updateQueryParameters.spec.ts
@@ -6,7 +6,7 @@ import { MockStore, provideMockStore } from "@ngrx/store/testing"
 import { waitFor } from "@testing-library/angular"
 import { Subject } from "rxjs"
 import { LoadInitialFileService } from "../../../features/loadFile/facade"
-import { EDGE_METRIC_DATA, METRIC_DATA } from "../../../util/dataMocks"
+import { EDGE_METRIC_DATA, METRIC_DATA } from "../../../mocks/dataMocks"
 import { metricDataSelector } from "../../selectors/accumulatedData/metricData/metricData.selector"
 import { setColorMetric } from "../../store/dynamicSettings/colorMetric/colorMetric.actions"
 import { setEdgeMetric } from "../../store/dynamicSettings/edgeMetric/edgeMetric.actions"
