+++ visualization/app/codeCharta/state/selectors/accumulatedData/metricData/edgeMetricData.calculator.spec.ts
@@ -1,5 +1,5 @@
 import { calculateEdgeMetricData, calculateNodePath } from "./edgeMetricData.calculator"
-import { FILE_STATES, FILE_STATES_TWO_FILES, VALID_NODE_WITH_PATH } from "../../../../util/dataMocks"
+import { FILE_STATES, FILE_STATES_TWO_FILES, VALID_NODE_WITH_PATH } from "../../../../mocks/dataMocks"
 import { FileState } from "../../../../model/files/files"
 import { clone } from "../../../../util/clone"
 import { createBlacklistMatcher } from "../../../../util/blacklist/blacklistMatcher"
