+++ visualization/app/codeCharta/state/selectors/accumulatedData/metricData/nodeMetricData.calculator.spec.ts
@@ -1,7 +1,7 @@
 import { FileSelectionState, FileState } from "../../../../model/files/files"
 import { clone } from "../../../../util/clone"
 import { createBlacklistMatcher } from "../../../../util/blacklist/blacklistMatcher"
-import { TEST_DELTA_MAP_A, VALID_NODE_WITH_ROOT_UNARY } from "../../../../util/dataMocks"
+import { TEST_DELTA_MAP_A, VALID_NODE_WITH_ROOT_UNARY } from "../../../../mocks/dataMocks"
 import { NodeDecorator } from "../../../../util/nodeDecorator"
 import { UNARY_METRIC, calculateNodeMetricData } from "./nodeMetricData.calculator"
