+++ visualization/app/codeCharta/state/selectors/amountOfBuildingsWithSelectedEdgeMetric/amountOfBuildingsWithSelectedEdgeMetric.selector.spec.ts
@@ -4,7 +4,7 @@ import { appReducers } from "../../store/state.manager"
 import { firstValueFrom } from "rxjs"
 import { amountOfBuildingsWithSelectedEdgeMetricSelector } from "./amountOfBuildingsWithSelectedEdgeMetric.selector"
 import { setFiles } from "../../store/files/files.actions"
-import { DEFAULT_STATE, FILE_STATES } from "../../../util/dataMocks"
+import { DEFAULT_STATE, FILE_STATES } from "../../../mocks/dataMocks"
 import { CcState } from "../../../codeCharta.model"
 import { setState } from "../../store/state.actions"
 import { setEdgeMetric } from "../../store/dynamicSettings/edgeMetric/edgeMetric.actions"
